Update shadow_attack.py
In the examples for generating attacks, all data is passed in at once, if this is the case then why is the data being resized to length of data * batch size? I cannot make this attack work, I have a feeling this is due to incorrect shaping of the data here.