Fix bug in PixelThreshold for single input
Calling `.generate()` on a batch of length one, a `TypeError` is raised on line 193. This is because `np.squeeze(y)` returns a 0-dimensional array when `y.shape == (1, 1)`. Changing it to `y.flatten()` solves this issue by returning a 1-dimensional array in this case, while not changing the behavior in the case when the batch has more than one input.