Create BacKBayDiffMod
This backdoor attack is a technique which implements a poisoning attack with a clean label backdoor. It contains methods such as `Poisoning Attack, (which takes as input the audio data and corresponding labels and returns the poisoned audio data and labels)` to apply the attack to the audio data, Bayesian style is implemented using a `prior and pymc framework with the Fokker-Planck equation for sampling` to obtain and define the prior distribution, and a diffusion technique is then applied to name: ``texttt{back\_diffusion\_sampling} (which implements a diffusion-based sampling technique to generate a sequence of samples as a function of certain parameters and a noise distribution.